Parsers/PythonParser/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
azure-core==1.32.0
azure-identity==1.21.0
azure-storage-blob==12.25.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozenlist==1.5.0
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
isodate==0.7.2
metar-taf-parser-mivek==1.9.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
propcache==0.3.0
protobuf==5.29.3
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pyftpdlib==2.0.1
PyJWT==2.10.1
requests==2.32.3
setuptools==76.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
